popup/views: reskin Welcome page (#352)
* welcome page reskin

* button spacing

* remove unused svg
