+++ extension/src/popup/views/Welcome/styles.scss
@@ -1,0 +1,7 @@
+.Welcome {
+  &--button-container {
+    display: flex;
+    justify-content: space-around;
+    padding-top: 1rem;
+  }
+}
